feat: hook useCreateAquarium
