+++ src/hooks/index.ts
@@ -1,2 +1,3 @@
 export { useAquariumDetails } from './useAquariumDetails';
+export { useCreateAquarium } from './useCreateAquarium';
 export { useResetParams } from './useResetParams';
