+++ src/routes/NewPets/index.tsx
@@ -1,40 +1,22 @@
-import { useState } from 'react';
-import { View, ScrollView, Alert } from 'react-native';
+import { View, ScrollView } from 'react-native';
 import { Text } from '@rneui/themed';
-import axios from 'axios';
 import { ItemButton, DefaultPrimaryButton, DefaultSecondaryButton, Loading } from '../../components';
-import { AquariumDataProps } from '../../types';
+import { useCreateAquarium } from '../../hooks';
 import { useAquarium } from '../../context';
 import { Icons } from '../../theme';
 import { S } from './styles';
 
-import { useResetParams } from '../../hooks';
-
 
 export default function NewPets({navigation}:any) {
 
-  const [ isLoading, setIsLoading ] = useState(false);
-  const resetParams = useResetParams();
+  const { createAquarium, loading } = useCreateAquarium();
 
   const {
-    token,
-    setAquariumsList,
-
-    aquariumName, selectedShape, selectedMaterial, selectedVoltage, thickness, height, volume, 
-    hasPump, hasFeeder, hasThermostat, hasFilter, hasLedLights, hasVegetation, 
-    hasTemperatureSensor, hasWaterLevelSensor, hasLuminositySensor, hasPhSensor, hasSaturationSensor, 
-
-    hasFish, setHasFish,
-    hasTurtle, setHasTurtle,
-    hasSnake, setHasSnake,
-    hasFrog, setHasFrog,
-    fishQuantity, setFishQuantity,
-    turtleQuantity, setTurtleQuantity,
-    snakeQuantity, setSnakeQuantity,
-    frogQuantity, setFrogQuantity,
+    hasFish, setHasFish, fishQuantity, setFishQuantity,
+    hasTurtle, setHasTurtle, turtleQuantity, setTurtleQuantity,
+    hasSnake, setHasSnake, snakeQuantity, setSnakeQuantity,
+    hasFrog, setHasFrog, frogQuantity, setFrogQuantity,
   } = useAquarium();
-
-  const headers = { 'Authorization': `${token}` };
 
   const pets = [
     { title: 'Peixe', icon: Icons.fish, onPress: () => setHasFish(!hasFish), isSelected: hasFish, itemQuantity: fishQuantity, setQuantity: setFishQuantity},
@@ -44,78 +26,13 @@ export default function NewPets({navigation}:any) {
   ]
 
   const handleCreateAquarium = async () => {
-
-    setIsLoading(true);
-
-    let aquariumData: AquariumDataProps = {
-      name: aquariumName, 
-      format_aquarium: selectedShape,
-      material: selectedMaterial,
-      voltage: selectedVoltage,
-      thickness: `${thickness}`,
-      height: `${height}`,
-      capacity: `${volume}`,
-    };
-
-    const aquariumAccessories = [
-      hasPump && { name: 'Bombinha' },
-      hasFeeder && { name: 'Alimentador automático' },
-      hasThermostat && { name: 'Termostato / Aquecedor' },
-      hasFilter && { name: 'Filtro' },
-      hasLedLights && { name: 'Luz LED' },
-      hasVegetation && { name: 'Plantas naturais' },
-    ].filter(Boolean);
-
-    const aquariumSensors = [
-      hasTemperatureSensor && { name: 'Sensor', metric: 'Temperatura', current: '27°C' },
-      hasWaterLevelSensor && { name: 'Sensor', metric: 'Nível de água', current: '90%' },
-      hasLuminositySensor && { name: 'Sensor', metric: 'Luminosidade', current: '35 lm'},
-      hasPhSensor && { name: 'Sensor', metric: 'pH', current: '7' },
-      hasSaturationSensor && { name: 'Sensor', metric: 'Saturação', current: '9,07 ppm'},
-    ].filter(Boolean);
-
-    const aquariumPets = [
-      hasFish && { species: 'Peixe', quantity: fishQuantity },
-      hasTurtle && { species: 'Tartaruga', quantity: turtleQuantity },
-      hasSnake && { species: 'Cobra', quantity: snakeQuantity },
-      hasFrog && { species: 'Sapo', quantity: frogQuantity },
-    ].filter(Boolean);
-
-    console.log(`Aquarium Data: ${JSON.stringify(aquariumData)}`);
-    console.log(`Aquarium Accessories: ${JSON.stringify(aquariumAccessories)}`);
-    console.log(`Aquarium Sensors: ${JSON.stringify(aquariumSensors)}`);
-    console.log(`Aquarium Pets: ${JSON.stringify(aquariumPets)}`);
-
-    try {
-      const aquariumResponse = await axios.post(`${process.env.BASE_URL}/aquarium`, aquariumData, { headers });
-      const aquariumId = aquariumResponse.data.result.id;
-
-      console.log(`Aquário criado com sucesso - ID: ${aquariumId}`);
-      aquariumData = { ...aquariumData, id: aquariumId };
-      setAquariumsList((prevList:AquariumDataProps[]) => [...prevList, aquariumData]);
-
-      await Promise.all([
-        ...aquariumAccessories.map(accessory => axios.post(`${process.env.BASE_URL}/aquarium/${aquariumId}/accessories`, accessory, { headers })),
-        ...aquariumSensors.map(sensor => axios.post(`${process.env.BASE_URL}/aquarium/${aquariumId}/sensors`, sensor, { headers })),
-        ...aquariumPets.map(pet => axios.post(`${process.env.BASE_URL}/aquarium/${aquariumId}/pets`, pet, { headers }))
-      ]);
-
-      console.log('Acessórios, sensores e pets adicionados com sucesso.');
-
-      resetParams();
-      navigation.navigate('AquariumsSelection');
-    } catch (e) {
-      Alert.alert('Erro', 'Houve um erro ao criar o aquário. Tente novamente mais tarde.')
-      console.error('Erro ao criar aquário:', e);
-    } finally {
-      setIsLoading(false);
-    }
+    createAquarium(navigation);
   };
 
   return (
     <ScrollView>
       <View style={S.page}>
-        {isLoading ? (
+        {loading ? (
           <Loading text='Criando aquário...' />
         ) : (
           <>
